Update patch set 6
Patch Set 6:

Arbri: PLEASE stop using Gerrit as Facebook. I've said it elsewhere.

Unless you have these devices and have tested this commit (impossible because it is broken) please don't vote on them.

CR+1 implies "we want this merged and it is of sufficient quality"

V+1 indicates "it works and builds"

Patch-set: 6